+++ cpp_conan/conanfile.py
@@ -10,7 +10,7 @@ class ExampleConan(ConanFile):
     settings = "os", "arch", "compiler", "build_type"
 
     requires = "boost/1.83.0"
-    tool_requires = "doxygen/1.9.4"
+    tool_requires = "doxygen/1.9.4", "cppcheck/2.12.1"
     test_requires = "gtest/1.14.0"
 
     generators = "CMakeDeps", "CMakeToolchain"