+++ cpp_conan/CHANGELOG.md
@@ -12,4 +12,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * doxygen documentation
 * example executable
 * linting with clang-tidy
+* linting with cppcheck
 * unit test